requirements.txt
# local package
-e .

# external requirements
click
Sphinx
coverage
python-dotenv>=0.5.1
Pillow